WIP: Started interpolation refactor.
